按照 review 要求更改了部分设计 - mapper 和 service 部分类和方法重命名 - 给原本的 mapper 做了了一个 CacheUrlMapperDecorator 装饰器增加了 redis 缓存查询和更新的功能 - 终于用上了 redisson, 愉快地把自己写的垃圾 BloomFilter 删了, 效率大幅度上升 - 将一些没必要的动态配置写死在 Configuration 里面 - 重新生成 ShortUrl 的流程检查后貌似没有问题 - code 和 msg 整合到枚举类里面了, 没有写 data 类, 暂时用一个 String 类代替, 目前前端没有需要复杂信息的要求, 比如获得链接的时间, 描述, 等等...(等后面有这个要求的时候我再用一个新的子类添加其他信息)
/**
 * ░░░░░░░░░░░░░░░░░░░░░░░░▄░░
 * ░░░░░░░░░▐█░░░░░░░░░░░▄▀▒▌░
 * ░░░░░░░░▐▀▒█░░░░░░░░▄▀▒▒▒▐
 * ░░░░░░░▐▄▀▒▒▀▀▀▀▄▄▄▀▒▒▒▒▒▐
 * ░░░░░▄▄▀▒░▒▒▒▒▒▒▒▒▒█▒▒▄█▒▐
 * ░░░▄▀▒▒▒░░░▒▒▒░░░▒▒▒▀██▀▒▌
 * ░░▐▒▒▒▄▄▒▒▒▒░░░▒▒▒▒▒▒▒▀▄▒▒
 * ░░▌░░▌█▀▒▒▒▒▒▄▀█▄▒▒▒▒▒▒▒█▒▐
 * ░▐░░░▒▒▒▒▒▒▒▒▌██▀▒▒░░░▒▒▒▀▄
 * ░▌░▒▄██▄▒▒▒▒▒▒▒▒▒░░░░░░▒▒▒▒
 * ▀▒▀▐▄█▄█▌▄░▀▒▒░░░░░░░░░░▒▒▒
 */

+++ src/main/java/com/dawn/shortlink/domain/pojo/ShortUrlRequestBody.java
@@ -7,7 +7,7 @@ public class ShortUrlRequestBody {
     public long timeout;
 
 
-    public ShortUrlRequestBody(String originUrl, String description, int timeout) {
+    public ShortUrlRequestBody(String originUrl, String description, long timeout) {
         this.originUrl = originUrl;
         this.description = description;
         this.timeout = timeout;
